Refactor layakine rendering into modular renderer
