+++ apps/layakine/render/draw3d.js
@@ -1,0 +1,1425 @@
+import {
+  getQuadrantMetrics,
+  getLineSoundPoints,
+  getPolygonSoundPoints,
+  lerpPoint,
+} from './draw2d.js';
+import {
+  createAxisRotationContext,
+  rotatePointAroundAxisOnPlane,
+} from '../jati3dGeometry.js';
+
+function clamp(value, min, max) {
+  if (!Number.isFinite(value)) {
+    return min;
+  }
+  return Math.min(Math.max(value, min), max);
+}
+
+let ctxRef = null;
+let canvasRef = null;
+let paletteRef = { stroke: '#404040', segment: '#f4f4f4', first: '#f4f4f4' };
+
+function setRenderContext({ ctx, canvas, palette }) {
+  ctxRef = ctx;
+  canvasRef = canvas;
+  paletteRef = {
+    stroke: palette?.stroke ?? '#404040',
+    segment: palette?.segment ?? '#f4f4f4',
+    first: palette?.first ?? palette?.segment ?? '#f4f4f4',
+  };
+}
+
+function getStrokeColor() {
+  return paletteRef.stroke;
+}
+
+function getSegmentColor() {
+  return paletteRef.segment;
+}
+
+function getFirstSoundMarkerColor() {
+  return paletteRef.first;
+}
+
+function getQuadrantMetricsLocal(quadrant) {
+  return getQuadrantMetrics(canvasRef, quadrant);
+}
+
+function projectPointToIsometric(point, baseCenter, origin, scale, height = 0) {
+  const dx = (point.x - baseCenter.x) * scale;
+  const dy = (point.y - baseCenter.y) * scale;
+  const x = origin.x + dx - dy;
+  const y = origin.y + (dx + dy) * 0.5 - height;
+  return { x, y };
+}
+
+function projectPointIso3d(point, origin, scale, verticalScale = 1) {
+  const scaledX = point.x * scale;
+  const scaledZ = point.z * scale;
+  const scaledY = point.y * verticalScale;
+  return {
+    x: origin.x + scaledX - scaledZ,
+    y: origin.y + (scaledX + scaledZ) * 0.5 - scaledY,
+  };
+}
+
+function buildJatiCrossSection(view2d, radius) {
+  if (!view2d) {
+    return { points: [], minY: 0, maxY: 0 };
+  }
+
+  const normalizePoints = (points) => {
+    if (!points.length) {
+      return { points: [], minY: 0, maxY: 0 };
+    }
+    let minY = Number.POSITIVE_INFINITY;
+    let maxY = Number.NEGATIVE_INFINITY;
+    points.forEach((point) => {
+      if (point.y < minY) {
+        minY = point.y;
+      }
+      if (point.y > maxY) {
+        maxY = point.y;
+      }
+    });
+    if (!Number.isFinite(minY) || !Number.isFinite(maxY)) {
+      return { points, minY: 0, maxY: 0 };
+    }
+    const midY = (minY + maxY) / 2;
+    const normalized = points.map((point) => ({ x: point.x, y: point.y - midY }));
+    return { points: normalized, minY: minY - midY, maxY: maxY - midY };
+  };
+
+  if (view2d.shape === 'circle') {
+    const segments = 32;
+    const points = [];
+    for (let i = 0; i < segments; i += 1) {
+      const angle = (i / segments) * Math.PI * 2;
+      points.push({ x: radius * Math.cos(angle), y: radius * Math.sin(angle) });
+    }
+    return normalizePoints(points);
+  }
+
+  if (view2d.shape === 'line') {
+    const halfLength = radius;
+    const thickness = Math.max(radius * 0.28, 6);
+    return normalizePoints([
+      { x: -halfLength, y: -thickness },
+      { x: halfLength, y: -thickness },
+      { x: halfLength, y: thickness },
+      { x: -halfLength, y: thickness },
+    ]);
+  }
+
+  if (view2d.shape === 'polygon') {
+    const sides = Math.max(3, view2d.sides || 3);
+    const points = [];
+    const rotation = -Math.PI / 2;
+    for (let i = 0; i < sides; i += 1) {
+      const angle = rotation + (i * 2 * Math.PI) / sides;
+      points.push({ x: radius * Math.cos(angle), y: radius * Math.sin(angle) });
+    }
+    return normalizePoints(points);
+  }
+
+  return { points: [], minY: 0, maxY: 0 };
+}
+function drawGatiQuadrant3dInternal(config, elapsed) {
+  const ctx = ctxRef;
+  const canvas = canvasRef;
+  const { orientation, view2d, cycleDuration } = config;
+  if (!ctx || !canvas || !view2d) {
+    return;
+  }
+
+  const { offsetX, offsetY, width, height, center } = getQuadrantMetricsLocal(orientation);
+  const baseCenter = center;
+  const isoOrigin = {
+    x: offsetX + width * 0.54,
+    y: clamp(baseCenter.y + height * 0.22, offsetY, offsetY + height),
+  };
+  const scale = 0.82;
+  const strokeColor = getStrokeColor('gati');
+  const segmentColor = getSegmentColor('gati');
+  const baseRadius = Math.min(width, height) * 0.32;
+  const baseMarkerRadius = Math.max(3, canvas.width * 0.0045);
+  const surfaceHeight = Math.max(2, baseMarkerRadius * 0.6);
+  const markerHeight = surfaceHeight;
+  const eventHeight = surfaceHeight;
+
+  const project = (point, heightOffset = 0) =>
+    projectPointToIsometric(point, baseCenter, isoOrigin, scale, heightOffset);
+
+  const baseCorners = [
+    { x: baseCenter.x - baseRadius, y: baseCenter.y - baseRadius },
+    { x: baseCenter.x + baseRadius, y: baseCenter.y - baseRadius },
+    { x: baseCenter.x + baseRadius, y: baseCenter.y + baseRadius },
+    { x: baseCenter.x - baseRadius, y: baseCenter.y + baseRadius },
+  ];
+  const isoCorners = baseCorners.map((corner) => project(corner, 0));
+  const stationaryMarkerEnabled =
+    view2d?.soundMarkers?.mode === 'count' && view2d.soundMarkers.count <= 2;
+  const farTopIndex = isoCorners.reduce(
+    (best, corner, index) => (corner.y < isoCorners[best].y ? index : best),
+    0,
+  );
+  const farRightIndex = isoCorners.reduce(
+    (best, corner, index) => (corner.x > isoCorners[best].x ? index : best),
+    0,
+  );
+  const farTopCorner = baseCorners[farTopIndex];
+  const farRightCorner = baseCorners[farRightIndex];
+  const topEdgeMidpoint = lerpPoint(farTopCorner, farRightCorner, 0.5);
+  const radialMargin = 0.16;
+  const directionToTop = {
+    x: topEdgeMidpoint.x - baseCenter.x,
+    y: topEdgeMidpoint.y - baseCenter.y,
+  };
+  const directionLength = Math.hypot(directionToTop.x, directionToTop.y) || 1;
+  const directionUnit = {
+    x: directionToTop.x / directionLength,
+    y: directionToTop.y / directionLength,
+  };
+  const shapeRadius = directionLength * (1 - radialMargin);
+  const firstMarkerBasePoint = {
+    x: baseCenter.x + directionUnit.x * shapeRadius,
+    y: baseCenter.y + directionUnit.y * shapeRadius,
+  };
+  const oppositeAxisPoint = {
+    x: baseCenter.x - directionUnit.x * shapeRadius,
+    y: baseCenter.y - directionUnit.y * shapeRadius,
+  };
+  const orientationAngle = Math.atan2(
+    firstMarkerBasePoint.y - baseCenter.y,
+    firstMarkerBasePoint.x - baseCenter.x,
+  );
+  const stationaryBasePoint = stationaryMarkerEnabled ? firstMarkerBasePoint : null;
+
+  const drawMarker = (point, options = {}) => {
+    const {
+      height: heightOffset = markerHeight,
+      radius = baseMarkerRadius,
+      color = segmentColor,
+      stroke = strokeColor,
+      baseOpacity = 0.2,
+    } = options;
+    const top = project(point, heightOffset);
+    ctx.save();
+    if (baseOpacity > 0) {
+      ctx.globalAlpha = baseOpacity;
+      ctx.fillStyle = color;
+      ctx.beginPath();
+      ctx.arc(top.x, top.y, radius * 1.8, 0, Math.PI * 2);
+      ctx.fill();
+      ctx.globalAlpha = 1;
+    }
+
+    ctx.fillStyle = color;
+    ctx.beginPath();
+    ctx.arc(top.x, top.y, radius, 0, Math.PI * 2);
+    ctx.fill();
+
+    ctx.strokeStyle = stroke;
+    ctx.lineWidth = Math.max(1, radius * 0.45);
+    ctx.stroke();
+    ctx.restore();
+  };
+
+  const drawBasePlane = () => {
+    ctx.save();
+    ctx.beginPath();
+    ctx.moveTo(isoCorners[0].x, isoCorners[0].y);
+    for (let i = 1; i < isoCorners.length; i += 1) {
+      ctx.lineTo(isoCorners[i].x, isoCorners[i].y);
+    }
+    ctx.closePath();
+    ctx.fillStyle = 'rgba(20, 44, 44, 0.7)';
+    ctx.fill();
+    ctx.strokeStyle = 'rgba(255, 255, 255, 0.08)';
+    ctx.lineWidth = 1.5;
+    ctx.stroke();
+
+    ctx.strokeStyle = 'rgba(255, 255, 255, 0.06)';
+    ctx.beginPath();
+    ctx.moveTo(isoCorners[0].x, isoCorners[0].y);
+    ctx.lineTo(isoCorners[2].x, isoCorners[2].y);
+    ctx.moveTo(isoCorners[1].x, isoCorners[1].y);
+    ctx.lineTo(isoCorners[3].x, isoCorners[3].y);
+    ctx.stroke();
+    ctx.restore();
+  };
+
+  drawBasePlane();
+
+  const segmentDuration = view2d.segmentDuration || 0;
+  const segmentCount = view2d.segmentCount || 1;
+  const fallbackCycle = segmentDuration * Math.max(1, segmentCount);
+  const shapeCycle = cycleDuration > 0 ? cycleDuration : fallbackCycle;
+
+  const drawLineShape = () => {
+    const start = firstMarkerBasePoint;
+    const end = oppositeAxisPoint;
+    const isoStart = project(start, eventHeight);
+    const isoEnd = project(end, eventHeight);
+
+    ctx.save();
+    ctx.strokeStyle = strokeColor;
+    ctx.lineWidth = 3;
+    ctx.lineJoin = 'round';
+    ctx.beginPath();
+    ctx.moveTo(isoStart.x, isoStart.y);
+    ctx.lineTo(isoEnd.x, isoEnd.y);
+    ctx.stroke();
+    ctx.restore();
+
+    const eventPoints = getLineSoundPoints(start, end, view2d, view2d.soundMarkers);
+    if (stationaryBasePoint) {
+      if (eventPoints.length > 0) {
+        eventPoints[0] = stationaryBasePoint;
+      } else {
+        eventPoints.push(stationaryBasePoint);
+      }
+    }
+    eventPoints.forEach((pt) => {
+      drawMarker(pt, { height: eventHeight });
+    });
+
+    if (!(segmentDuration > 0)) {
+      const staticPoint = stationaryBasePoint || start;
+      drawMarker(staticPoint, {
+        height: eventHeight,
+        radius: baseMarkerRadius * 1.2,
+        baseOpacity: 0.28,
+      });
+      return;
+    }
+
+    let progressPoint = start;
+    if (view2d.bounce) {
+      const cycleDuration = segmentDuration * 2;
+      if (cycleDuration > 0) {
+        const local = elapsed % cycleDuration;
+        const index = Math.floor(local / segmentDuration);
+        const t = (local - index * segmentDuration) / segmentDuration;
+        progressPoint = index % 2 === 0 ? lerpPoint(start, end, t) : lerpPoint(end, start, t);
+      }
+    } else if (view2d.segmentCount && view2d.segmentCount > 1) {
+      const cycle = segmentDuration * view2d.segmentCount;
+      if (cycle > 0) {
+        const local = elapsed % cycle;
+        const index = Math.floor(local / segmentDuration);
+        const t = (local - index * segmentDuration) / segmentDuration;
+        const current = index % 2 === 0 ? start : end;
+        const next = index % 2 === 0 ? end : start;
+        progressPoint = lerpPoint(current, next, t);
+      }
+    } else {
+      const local = elapsed % segmentDuration;
+      const t = segmentDuration > 0 ? local / segmentDuration : 0;
+      progressPoint = lerpPoint(start, end, t);
+    }
+
+    drawMarker(progressPoint, {
+      height: eventHeight,
+      radius: baseMarkerRadius * 1.2,
+      baseOpacity: 0.32,
+    });
+  };
+
+  const drawPolygonShape = () => {
+    const sides = Math.max(3, view2d.sides || 3);
+    const points = [];
+    for (let i = 0; i < sides; i += 1) {
+      const angle = orientationAngle + (i * 2 * Math.PI) / sides;
+      points.push({
+        x: baseCenter.x + shapeRadius * Math.cos(angle),
+        y: baseCenter.y + shapeRadius * Math.sin(angle),
+      });
+    }
+
+    ctx.save();
+    ctx.strokeStyle = strokeColor;
+    ctx.lineWidth = 3;
+    ctx.beginPath();
+    ctx.moveTo(points[0].x, points[0].y);
+    for (let i = 1; i < points.length; i += 1) {
+      ctx.lineTo(points[i].x, points[i].y);
+    }
+    ctx.closePath();
+    ctx.stroke();
+    ctx.restore();
+
+    const markerPoints = getPolygonSoundPoints(points, view2d.soundMarkers);
+    markerPoints.forEach((point) => drawMarker(point, { height: eventHeight }));
+
+    if (!(segmentDuration > 0)) {
+      const staticPoint = points[0];
+      drawMarker(staticPoint, {
+        height: eventHeight,
+        radius: baseMarkerRadius * 1.2,
+        baseOpacity: 0.28,
+      });
+      return;
+    }
+
+    const cycle = segmentDuration * sides;
+    const local = cycle > 0 ? elapsed % cycle : 0;
+    const index = segmentDuration > 0 ? Math.floor(local / segmentDuration) : 0;
+    const t = segmentDuration > 0 ? (local - index * segmentDuration) / segmentDuration : 0;
+    const current = points[index % points.length];
+    const next = points[(index + 1) % points.length];
+    const progressPoint = lerpPoint(current, next, t);
+
+    drawMarker(progressPoint, {
+      height: eventHeight,
+      radius: baseMarkerRadius * 1.25,
+      baseOpacity: 0.32,
+    });
+  };
+
+  const drawCircleShape = () => {
+    const markers = (() => {
+      if (view2d?.soundMarkers?.mode === 'count') {
+        const count = Math.max(1, Math.floor(view2d.soundMarkers.count));
+        return Array.from({ length: count }, (_, index) => {
+          const angle = orientationAngle + (index / count) * Math.PI * 2;
+          return {
+            x: baseCenter.x + shapeRadius * Math.cos(angle),
+            y: baseCenter.y + shapeRadius * Math.sin(angle),
+          };
+        });
+      }
+      return [firstMarkerBasePoint];
+    })();
+
+    markers.forEach((point) => drawMarker(point, { height: eventHeight }));
+
+    const local = shapeCycle > 0 ? elapsed % shapeCycle : 0;
+    const progress = shapeCycle > 0 ? local / shapeCycle : 0;
+    const angle = orientationAngle + 2 * Math.PI * progress;
+    const progressPoint = {
+      x: baseCenter.x + shapeRadius * Math.cos(angle),
+      y: baseCenter.y + shapeRadius * Math.sin(angle),
+    };
+
+    drawMarker(progressPoint, {
+      height: eventHeight,
+      radius: baseMarkerRadius * 1.25,
+      baseOpacity: 0.32,
+    });
+  };
+
+  if (view2d.shape === 'line') {
+    drawLineShape();
+  } else if (view2d.shape === 'polygon') {
+    drawPolygonShape();
+  } else if (view2d.shape === 'circle') {
+    drawCircleShape();
+  }
+}
+      baseOpacity: 0.32,
+    });
+  };
+
+  if (view2d.shape === 'line') {
+    drawLineShape();
+  } else if (view2d.shape === 'polygon') {
+    drawPolygonShape();
+  } else if (view2d.shape === 'circle') {
+    drawCircleShape();
+  }
+}
+
+function drawJatiQuadrant3dInternal(config, elapsed) {
+  const ctx = ctxRef;
+  const canvas = canvasRef;
+
+  const { orientation, view2d, cycleDuration, gatiCount: rawGatiCount = 1 } = config;
+  if (!ctx || !canvas || !view2d) {
+    return;
+  }
+
+  const { offsetX, offsetY, width, height, center, verticalShift } =
+    getQuadrantMetricsLocal(orientation);
+  const origin = {
+    x: offsetX + width * 0.54,
+    y: clamp(offsetY + height * 0.68 + verticalShift, offsetY, offsetY + height),
+  };
+  const baseRadius = Math.min(width, height) * 0.22;
+  const shapeRadius = baseRadius * 0.55;
+  const crossSectionData = buildJatiCrossSection(view2d, shapeRadius);
+  const shapePoints = crossSectionData.points;
+  if (shapePoints.length < 2) {
+    return;
+  }
+
+  const strokeColor = getStrokeColor('jati');
+  const segmentColor = getSegmentColor('jati');
+  const scale = 1;
+  const verticalScale = 0.9;
+  const baseLift = crossSectionData.maxY;
+  const isLineShape = view2d.shape === 'line';
+  const isCircleShape = view2d.shape === 'circle';
+  const minLocalX = shapePoints.reduce(
+    (minX, point) => (point.x < minX ? point.x : minX),
+    Number.POSITIVE_INFINITY,
+  );
+  const maxLocalX = shapePoints.reduce(
+    (maxX, point) => (point.x > maxX ? point.x : maxX),
+    Number.NEGATIVE_INFINITY,
+  );
+  const gatiCount = Math.max(1, Math.floor(rawGatiCount) || 1);
+  const copyCount = gatiCount;
+  const showFullScene = true;
+
+  const segmentDuration = view2d.segmentDuration || 0;
+  const cycleSegments = view2d.segmentCount || 1;
+  const fallbackCycle = segmentDuration * Math.max(1, cycleSegments);
+  const copyCycle = cycleDuration > 0 ? cycleDuration : fallbackCycle;
+  const totalCycle = copyCycle > 0 ? copyCycle * copyCount : 0;
+
+  let activeCopyIndex = 0;
+  let copyElapsed = 0;
+  if (copyCycle > 0 && totalCycle > 0) {
+    const wrapped = ((elapsed % totalCycle) + totalCycle) % totalCycle;
+    activeCopyIndex = Math.floor(wrapped / copyCycle) % copyCount;
+    copyElapsed = wrapped - activeCopyIndex * copyCycle;
+  }
+
+  const baseAngle = -Math.PI / 2;
+  const angleStep = copyCount > 0 ? (Math.PI * 2) / copyCount : 0;
+
+  const innerPointIndex = shapePoints.reduce(
+    (bestIndex, point, index) => (point.x < shapePoints[bestIndex].x ? index : bestIndex),
+    0,
+  );
+  const epsilon = 1e-6;
+  const innerPointLocalX = shapePoints[innerPointIndex]?.x ?? 0;
+  const innerPointCandidates = shapePoints
+    .map((pt, index) => ({ pt, index }))
+    .filter(({ pt }) => Math.abs(pt.x - innerPointLocalX) < epsilon);
+
+  const primaryInnerCandidate = (() => {
+    if (!innerPointCandidates.length) {
+      return null;
+    }
+    const directMatch = innerPointCandidates.find(({ index }) => index === innerPointIndex);
+    if (directMatch) {
+      return directMatch;
+    }
+    return innerPointCandidates.reduce((best, candidate) =>
+      candidate.pt.y > best.pt.y ? candidate : best,
+    innerPointCandidates[0]);
+  })();
+
+  const referenceInnerPoint = primaryInnerCandidate?.pt ?? shapePoints[innerPointIndex] ?? {
+    x: innerPointLocalX,
+    y: 0,
+  };
+  const referenceInnerIndex = primaryInnerCandidate?.index ?? innerPointIndex;
+
+  const soundCircleRadius = baseRadius + innerPointLocalX;
+  const soundWorldPoint = {
+    x: soundCircleRadius * Math.cos(baseAngle),
+    y: baseLift - referenceInnerPoint.y,
+    z: soundCircleRadius * Math.sin(baseAngle),
+  };
+  let soundIsoPoint = projectPointIso3d(soundWorldPoint, origin, scale, verticalScale);
+
+  const baseMarkerRadius = Math.max(4, canvas.width * 0.0046);
+
+  const drawMarker = (point, options = {}) => {
+    const {
+      color = segmentColor,
+      stroke = strokeColor,
+      baseOpacity = 0.25,
+      radius = baseMarkerRadius,
+    } = options;
+    ctx.save();
+    if (baseOpacity > 0) {
+      ctx.globalAlpha = baseOpacity;
+      ctx.fillStyle = color;
+      ctx.beginPath();
+      ctx.arc(point.x, point.y, radius * 1.7, 0, Math.PI * 2);
+      ctx.fill();
+      ctx.globalAlpha = 1;
+    }
+    ctx.fillStyle = color;
+    ctx.beginPath();
+    ctx.arc(point.x, point.y, radius, 0, Math.PI * 2);
+    ctx.fill();
+    ctx.strokeStyle = stroke;
+    ctx.lineWidth = Math.max(1, radius * 0.35);
+    ctx.stroke();
+    ctx.restore();
+  };
+
+  const copyInfos = Array.from({ length: copyCount }, (_, index) => {
+    const angle = baseAngle + angleStep * index;
+    const cos = Math.cos(angle);
+    const sin = Math.sin(angle);
+    const basePoint = { x: cos * baseRadius, y: 0, z: sin * baseRadius };
+    const isoBase = projectPointIso3d(basePoint, origin, scale, verticalScale);
+
+    const points = shapePoints.map((pt, pointIndex) => {
+      const radialDistance = baseRadius + pt.x;
+      const worldPoint = {
+        x: radialDistance * cos,
+        y: baseLift - pt.y,
+        z: radialDistance * sin,
+      };
+      const isoPoint = projectPointIso3d(worldPoint, origin, scale, verticalScale);
+      return {
+        pointIndex,
+        local: pt,
+        world: worldPoint,
+        iso: isoPoint,
+      };
+    });
+
+    const innerPointIsos = points.filter((pt) => Math.abs(pt.local.x - innerPointLocalX) < epsilon);
+    let referenceIso = null;
+    const directReference = points.find((pt) => pt.pointIndex === referenceInnerIndex);
+    if (directReference) {
+      referenceIso = directReference.iso;
+    } else if (innerPointIsos.length > 0) {
+      const sum = innerPointIsos.reduce(
+        (acc, pt) => ({ x: acc.x + pt.iso.x, y: acc.y + pt.iso.y }),
+        { x: 0, y: 0 },
+      );
+      referenceIso = { x: sum.x / innerPointIsos.length, y: sum.y / innerPointIsos.length };
+    } else if (points[innerPointIndex]) {
+      referenceIso = points[innerPointIndex].iso;
+    }
+
+    const offset = referenceIso
+      ? {
+          x: soundIsoPoint.x - referenceIso.x,
+          y: soundIsoPoint.y - referenceIso.y,
+        }
+      : { x: 0, y: 0 };
+
+    const translatedPoints = points.map((pt) => ({
+      ...pt,
+      iso: { x: pt.iso.x + offset.x, y: pt.iso.y + offset.y },
+    }));
+    const translatedIsoBase = { x: isoBase.x + offset.x, y: isoBase.y + offset.y };
+
+    const pointCount = translatedPoints.length;
+    const pathIsoPoints = [];
+    for (let i = 0; i <= pointCount; i += 1) {
+      const point = translatedPoints[(innerPointIndex + i) % pointCount];
+      pathIsoPoints.push(point.iso);
+    }
+
+    const pathSegmentLengths = [];
+    let pathTotalLength = 0;
+    for (let i = 0; i < pathIsoPoints.length - 1; i += 1) {
+      const from = pathIsoPoints[i];
+      const to = pathIsoPoints[i + 1];
+      const length = Math.hypot(to.x - from.x, to.y - from.y);
+      pathSegmentLengths.push(length);
+      pathTotalLength += length;
+    }
+
+    const radialTarget = translatedPoints.reduce((best, pt) => {
+      const bestDist = Math.hypot(best.iso.x - soundIsoPoint.x, best.iso.y - soundIsoPoint.y);
+      const dist = Math.hypot(pt.iso.x - soundIsoPoint.x, pt.iso.y - soundIsoPoint.y);
+      return dist > bestDist ? pt : best;
+    }, translatedPoints[0]);
+
+    let lineSegment = null;
+    if (isLineShape && Number.isFinite(minLocalX) && Number.isFinite(maxLocalX)) {
+      const epsilon = 1e-6;
+      const startCandidates = translatedPoints.filter(
+        (pt) => Math.abs(pt.local.x - minLocalX) < epsilon,
+      );
+      const endCandidates = translatedPoints.filter(
+        (pt) => Math.abs(pt.local.x - maxLocalX) < epsilon,
+      );
+      const averageIso = (candidates) => {
+        if (!candidates.length) {
+          return null;
+        }
+        const sum = candidates.reduce(
+          (acc, pt) => ({ x: acc.x + pt.iso.x, y: acc.y + pt.iso.y }),
+          { x: 0, y: 0 },
+        );
+        return { x: sum.x / candidates.length, y: sum.y / candidates.length };
+      };
+      const isoStart = averageIso(startCandidates);
+      const isoEnd = averageIso(endCandidates);
+      if (isoStart && isoEnd) {
+        const lineLength = Math.hypot(isoEnd.x - isoStart.x, isoEnd.y - isoStart.y);
+        lineSegment = { start: isoStart, end: isoEnd, length: lineLength };
+      }
+    }
+
+    if (lineSegment) {
+      return {
+        index,
+        angle,
+        cos,
+        sin,
+        depth: basePoint.z,
+        facing: cos >= 0,
+        isoBase: translatedIsoBase,
+        points: translatedPoints,
+        pathIsoPoints: [lineSegment.start, lineSegment.end],
+        pathSegmentLengths: [lineSegment.length],
+        pathTotalLength: lineSegment.length,
+        radialIso: lineSegment.end || translatedIsoBase,
+        lineSegment,
+      };
+    }
+
+    return {
+      index,
+      angle,
+      cos,
+      sin,
+      depth: basePoint.z,
+      facing: cos >= 0,
+      isoBase: translatedIsoBase,
+      points: translatedPoints,
+      pathIsoPoints,
+      pathSegmentLengths,
+      pathTotalLength,
+      radialIso: radialTarget?.iso || translatedIsoBase,
+      lineSegment: null,
+    };
+  });
+
+  const bounds = {
+    minX: Number.POSITIVE_INFINITY,
+    minY: Number.POSITIVE_INFINITY,
+    maxX: Number.NEGATIVE_INFINITY,
+    maxY: Number.NEGATIVE_INFINITY,
+  };
+
+  const includePointInBounds = (point) => {
+    if (!point || !Number.isFinite(point.x) || !Number.isFinite(point.y)) {
+      return;
+    }
+    if (point.x < bounds.minX) bounds.minX = point.x;
+    if (point.y < bounds.minY) bounds.minY = point.y;
+    if (point.x > bounds.maxX) bounds.maxX = point.x;
+    if (point.y > bounds.maxY) bounds.maxY = point.y;
+  };
+
+  includePointInBounds(soundIsoPoint);
+  copyInfos.forEach((info) => {
+    includePointInBounds(info.isoBase);
+    includePointInBounds(info.radialIso);
+    info.pathIsoPoints.forEach(includePointInBounds);
+    info.points.forEach((pt) => includePointInBounds(pt.iso));
+    if (info.lineSegment) {
+      includePointInBounds(info.lineSegment.start);
+      includePointInBounds(info.lineSegment.end);
+    }
+  });
+
+  if (
+    Number.isFinite(bounds.minX) &&
+    Number.isFinite(bounds.minY) &&
+    Number.isFinite(bounds.maxX) &&
+    Number.isFinite(bounds.maxY)
+  ) {
+    const currentCenter = {
+      x: (bounds.minX + bounds.maxX) / 2,
+      y: (bounds.minY + bounds.maxY) / 2,
+    };
+    const targetCenter = center;
+    const shift = {
+      x: targetCenter.x - currentCenter.x,
+      y: targetCenter.y - currentCenter.y,
+    };
+
+    if (Math.abs(shift.x) > 1e-6 || Math.abs(shift.y) > 1e-6) {
+      const translatePoint = (point) =>
+        point
+          ? {
+              x: point.x + shift.x,
+              y: point.y + shift.y,
+            }
+          : point;
+
+      soundIsoPoint = translatePoint(soundIsoPoint);
+      copyInfos.forEach((info) => {
+        info.isoBase = translatePoint(info.isoBase);
+        info.radialIso = translatePoint(info.radialIso);
+        info.pathIsoPoints = info.pathIsoPoints.map(translatePoint);
+        info.points = info.points.map((pt) => ({
+          ...pt,
+          iso: translatePoint(pt.iso),
+        }));
+        if (info.lineSegment) {
+          info.lineSegment = {
+            ...info.lineSegment,
+            start: translatePoint(info.lineSegment.start),
+            end: translatePoint(info.lineSegment.end),
+          };
+        }
+      });
+    }
+  }
+
+  const drawInfos = [...copyInfos].sort((a, b) => a.depth - b.depth);
+
+  const drawShape = (info, options = {}) => {
+    const { activeCopyIndex: activeIndex } = options;
+    const isoPoints = info.points.map((pt) => pt.iso);
+    if (!isoPoints.length) {
+      return;
+    }
+    const isActive = info.index === activeIndex;
+    const frontShade = info.facing ? 0.26 : 0.14;
+    const baseAlpha = showFullScene ? frontShade : frontShade + 0.12;
+    const strokeAlpha = isActive ? 0.9 : info.facing ? 0.65 : 0.4;
+    const highlightFirstCopy = info.index === 0;
+    const highlightStrokeStyle = 'rgba(255, 255, 255, 0.23)';
+    const highlightLineWidth = Math.max(0.7, canvas.width * 0.001);
+    if (isLineShape && info.lineSegment) {
+      ctx.save();
+      ctx.strokeStyle = `rgba(93, 42, 44, ${strokeAlpha})`;
+      ctx.lineWidth = Math.max(1.8, canvas.width * (isActive ? 0.0028 : 0.002));
+      ctx.beginPath();
+      ctx.moveTo(info.lineSegment.start.x, info.lineSegment.start.y);
+      ctx.lineTo(info.lineSegment.end.x, info.lineSegment.end.y);
+      ctx.stroke();
+      if (highlightFirstCopy) {
+        ctx.strokeStyle = highlightStrokeStyle;
+        ctx.lineWidth = highlightLineWidth;
+        ctx.stroke();
+      }
+      ctx.restore();
+      return;
+    }
+    if (isCircleShape) {
+      ctx.save();
+      ctx.strokeStyle = `rgba(93, 42, 44, ${strokeAlpha})`;
+      ctx.lineWidth = Math.max(1.8, canvas.width * (isActive ? 0.0026 : 0.0019));
+      ctx.beginPath();
+      ctx.moveTo(isoPoints[0].x, isoPoints[0].y);
+      for (let i = 1; i < isoPoints.length; i += 1) {
+        ctx.lineTo(isoPoints[i].x, isoPoints[i].y);
+      }
+      ctx.closePath();
+      ctx.stroke();
+      if (highlightFirstCopy) {
+        ctx.strokeStyle = highlightStrokeStyle;
+        ctx.lineWidth = highlightLineWidth;
+        ctx.stroke();
+      }
+      ctx.restore();
+      return;
+    }
+    ctx.save();
+    ctx.beginPath();
+    ctx.moveTo(isoPoints[0].x, isoPoints[0].y);
+    for (let i = 1; i < isoPoints.length; i += 1) {
+      ctx.lineTo(isoPoints[i].x, isoPoints[i].y);
+    }
+    ctx.closePath();
+    ctx.fillStyle = `rgba(231, 111, 81, ${Math.min(0.85, baseAlpha + (isActive ? 0.18 : 0))})`;
+    ctx.fill();
+    ctx.strokeStyle = `rgba(93, 42, 44, ${strokeAlpha})`;
+    ctx.lineWidth = Math.max(1.6, canvas.width * (isActive ? 0.0024 : 0.0016));
+    ctx.stroke();
+    if (highlightFirstCopy) {
+      ctx.strokeStyle = highlightStrokeStyle;
+      ctx.lineWidth = highlightLineWidth;
+      ctx.stroke();
+    }
+    ctx.restore();
+  };
+
+  let movingIsoPoint = null;
+  if (copyCycle > 0 && copyInfos[activeCopyIndex]) {
+    const info = copyInfos[activeCopyIndex];
+    const totalLength = info.pathTotalLength;
+    const segmentDuration = view2d.segmentDuration || 0;
+    if (view2d.bounce && info.lineSegment && segmentDuration > 0) {
+      const bounceCycle = segmentDuration * 2;
+      const local = bounceCycle > 0 ? copyElapsed % bounceCycle : 0;
+      const phase = bounceCycle > 0 ? Math.floor(local / segmentDuration) : 0;
+      const t = segmentDuration > 0 ? (local - phase * segmentDuration) / segmentDuration : 0;
+      const fromPoint = phase % 2 === 0 ? info.lineSegment.start : info.lineSegment.end;
+      const toPoint = phase % 2 === 0 ? info.lineSegment.end : info.lineSegment.start;
+      movingIsoPoint = {
+        x: fromPoint.x + (toPoint.x - fromPoint.x) * t,
+        y: fromPoint.y + (toPoint.y - fromPoint.y) * t,
+      };
+    } else if (totalLength > 0) {
+      const progress = copyElapsed / copyCycle;
+      let distance = Math.max(0, Math.min(1, progress)) * totalLength;
+      for (let i = 0; i < info.pathSegmentLengths.length; i += 1) {
+        const length = info.pathSegmentLengths[i];
+        const from = info.pathIsoPoints[i];
+        const to = info.pathIsoPoints[i + 1];
+        if (!(length > 0)) {
+          continue;
+        }
+        if (distance <= length) {
+          const tSegment = distance / length;
+          movingIsoPoint = {
+            x: from.x + (to.x - from.x) * tSegment,
+            y: from.y + (to.y - from.y) * tSegment,
+          };
+          break;
+        }
+        distance -= length;
+      }
+      if (!movingIsoPoint) {
+        const last = info.pathIsoPoints[info.pathIsoPoints.length - 1];
+        movingIsoPoint = last || info.pathIsoPoints[0];
+      }
+    }
+  }
+
+  if (!movingIsoPoint && copyInfos[0]) {
+    movingIsoPoint = copyInfos[0].pathIsoPoints[0];
+  }
+
+  drawInfos.forEach((info) => {
+    const shouldDraw = showFullScene || info.index === activeCopyIndex;
+    if (!shouldDraw) {
+      return;
+    }
+    drawShape(info, { activeCopyIndex });
+  });
+
+  if (movingIsoPoint) {
+    drawMarker(movingIsoPoint, {
+      radius: baseMarkerRadius * 1.1,
+      baseOpacity: 0.32,
+    });
+  }
+
+  drawMarker(soundIsoPoint, {
+    color: 'rgba(255, 255, 255, 0.85)',
+    stroke: 'rgba(255, 255, 255, 0.9)',
+    baseOpacity: 0.18,
+    radius: baseMarkerRadius * 0.75,
+    radius: baseMarkerRadius * 0.75,
+  });
+}
+
+function drawNadaiQuadrant3dInternal(config, elapsed) {
+  const ctx = ctxRef;
+  const canvas = canvasRef;
+
+  const { orientation, view2d, cycleDuration, gatiCount: rawGatiCount = 1 } = config;
+  if (!view2d) {
+    return;
+  }
+
+  const { offsetX, offsetY, width, height, center, verticalShift } =
+    getQuadrantMetricsLocal(orientation);
+  const origin = {
+    x: offsetX + width * 0.54,
+    y: clamp(offsetY + height * 0.68 + verticalShift, offsetY, offsetY + height),
+  };
+  const baseRadius = Math.min(width, height) * 0.22;
+  const shapeRadius = baseRadius * 0.55;
+  const crossSectionData = buildJatiCrossSection(view2d, shapeRadius);
+  const shapePoints = crossSectionData.points;
+  if (shapePoints.length < 2) {
+    return;
+  }
+
+  const strokeColor = getStrokeColor('nadai');
+  const segmentColor = getSegmentColor('nadai');
+  const scale = 1;
+  const verticalScale = 0.9;
+  const baseLift = crossSectionData.maxY;
+  const isLineShape = view2d.shape === 'line';
+  const isCircleShape = view2d.shape === 'circle';
+  const minLocalX = shapePoints.reduce(
+    (minX, point) => (point.x < minX ? point.x : minX),
+    Number.POSITIVE_INFINITY,
+  );
+  const maxLocalX = shapePoints.reduce(
+    (maxX, point) => (point.x > maxX ? point.x : maxX),
+    Number.NEGATIVE_INFINITY,
+  );
+  const gatiCount = Math.max(1, Math.floor(rawGatiCount) || 1);
+  const copyCount = gatiCount;
+  const showFullScene = true;
+
+  const segmentDuration = view2d.segmentDuration || 0;
+  const cycleSegments = view2d.segmentCount || 1;
+  const fallbackCycle = segmentDuration * Math.max(1, cycleSegments);
+  const copyCycle = fallbackCycle > 0 ? fallbackCycle : cycleDuration > 0 ? cycleDuration : 0;
+  const totalCycle = copyCycle > 0 ? copyCycle * copyCount : 0;
+
+  let activeCopyIndex = 0;
+  let copyElapsed = 0;
+  if (copyCycle > 0 && totalCycle > 0) {
+    const wrapped = ((elapsed % totalCycle) + totalCycle) % totalCycle;
+    activeCopyIndex = Math.floor(wrapped / copyCycle) % copyCount;
+    copyElapsed = wrapped - activeCopyIndex * copyCycle;
+  }
+
+  const baseAngle = -Math.PI / 2;
+  const angleStep = copyCount > 0 ? (Math.PI * 2) / copyCount : 0;
+
+  const innerPointIndex = shapePoints.reduce(
+    (bestIndex, point, index) => (point.x < shapePoints[bestIndex].x ? index : bestIndex),
+    0,
+  );
+  const epsilon = 1e-6;
+  const innerPointLocalX = shapePoints[innerPointIndex]?.x ?? 0;
+  const innerPointCandidates = shapePoints
+    .map((pt, index) => ({ pt, index }))
+    .filter(({ pt }) => Math.abs(pt.x - innerPointLocalX) < epsilon);
+
+  const primaryInnerCandidate = (() => {
+    if (!innerPointCandidates.length) {
+      return null;
+    }
+    const directMatch = innerPointCandidates.find(({ index }) => index === innerPointIndex);
+    if (directMatch) {
+      return directMatch;
+    }
+    return innerPointCandidates.reduce((best, candidate) =>
+      candidate.pt.y > best.pt.y ? candidate : best,
+    innerPointCandidates[0]);
+  })();
+
+  const referenceInnerPoint = primaryInnerCandidate?.pt ?? shapePoints[innerPointIndex] ?? {
+    x: innerPointLocalX,
+    y: 0,
+  };
+  const referenceInnerIndex = primaryInnerCandidate?.index ?? innerPointIndex;
+
+  const soundCircleRadius = baseRadius + innerPointLocalX;
+  const soundWorldPoint = {
+    x: soundCircleRadius * Math.cos(baseAngle),
+    y: baseLift - referenceInnerPoint.y,
+    z: soundCircleRadius * Math.sin(baseAngle),
+  };
+  let soundIsoPoint = projectPointIso3d(soundWorldPoint, origin, scale, verticalScale);
+
+  const baseMarkerRadius = Math.max(4, canvas.width * 0.0046);
+
+  const drawMarker = (point, options = {}) => {
+    const {
+      color = segmentColor,
+      stroke = strokeColor,
+      baseOpacity = 0.25,
+      radius = baseMarkerRadius,
+    } = options;
+    ctx.save();
+    if (baseOpacity > 0) {
+      ctx.globalAlpha = baseOpacity;
+      ctx.fillStyle = color;
+      ctx.beginPath();
+      ctx.arc(point.x, point.y, radius * 1.7, 0, Math.PI * 2);
+      ctx.fill();
+      ctx.globalAlpha = 1;
+    }
+    ctx.fillStyle = color;
+    ctx.beginPath();
+    ctx.arc(point.x, point.y, radius, 0, Math.PI * 2);
+    ctx.fill();
+    ctx.strokeStyle = stroke;
+    ctx.lineWidth = Math.max(1, radius * 0.35);
+    ctx.stroke();
+    ctx.restore();
+  };
+
+  const copyInfos = Array.from({ length: copyCount }, (_, index) => {
+    const angle = baseAngle + angleStep * index;
+    const cos = Math.cos(angle);
+    const sin = Math.sin(angle);
+    const basePoint = { x: cos * baseRadius, y: 0, z: sin * baseRadius };
+    const isoBase = projectPointIso3d(basePoint, origin, scale, verticalScale);
+
+    const points = shapePoints.map((pt, pointIndex) => {
+      const radialDistance = baseRadius + pt.x;
+      const worldPoint = {
+        x: radialDistance * cos,
+        y: baseLift - pt.y,
+        z: radialDistance * sin,
+      };
+      const isoPoint = projectPointIso3d(worldPoint, origin, scale, verticalScale);
+      return {
+        pointIndex,
+        local: pt,
+        world: worldPoint,
+        iso: isoPoint,
+      };
+    });
+
+    const innerPointIsos = points.filter((pt) => Math.abs(pt.local.x - innerPointLocalX) < epsilon);
+    let referenceIso = null;
+    const directReference = points.find((pt) => pt.pointIndex === referenceInnerIndex);
+    if (directReference) {
+      referenceIso = directReference.iso;
+    } else if (innerPointIsos.length > 0) {
+      const sum = innerPointIsos.reduce(
+        (acc, pt) => ({ x: acc.x + pt.iso.x, y: acc.y + pt.iso.y }),
+        { x: 0, y: 0 },
+      );
+      referenceIso = { x: sum.x / innerPointIsos.length, y: sum.y / innerPointIsos.length };
+    } else if (points[innerPointIndex]) {
+      referenceIso = points[innerPointIndex].iso;
+    }
+
+    const offset = referenceIso
+      ? {
+          x: soundIsoPoint.x - referenceIso.x,
+          y: soundIsoPoint.y - referenceIso.y,
+        }
+      : { x: 0, y: 0 };
+
+    const translatedPoints = points.map((pt) => ({
+      ...pt,
+      iso: { x: pt.iso.x + offset.x, y: pt.iso.y + offset.y },
+    }));
+    const translatedIsoBase = { x: isoBase.x + offset.x, y: isoBase.y + offset.y };
+
+    const pointCount = translatedPoints.length;
+    const pathIsoPoints = [];
+    for (let i = 0; i <= pointCount; i += 1) {
+      const point = translatedPoints[(innerPointIndex + i) % pointCount];
+      pathIsoPoints.push(point.iso);
+    }
+
+    const pathSegmentLengths = [];
+    let pathTotalLength = 0;
+    for (let i = 0; i < pathIsoPoints.length - 1; i += 1) {
+      const from = pathIsoPoints[i];
+      const to = pathIsoPoints[i + 1];
+      const length = Math.hypot(to.x - from.x, to.y - from.y);
+      pathSegmentLengths.push(length);
+      pathTotalLength += length;
+    }
+
+    const radialTarget = translatedPoints.reduce((best, pt) => {
+      const bestDist = Math.hypot(best.iso.x - soundIsoPoint.x, best.iso.y - soundIsoPoint.y);
+      const dist = Math.hypot(pt.iso.x - soundIsoPoint.x, pt.iso.y - soundIsoPoint.y);
+      return dist > bestDist ? pt : best;
+    }, translatedPoints[0]);
+
+    let lineSegment = null;
+    if (isLineShape && Number.isFinite(minLocalX) && Number.isFinite(maxLocalX)) {
+      const epsilon = 1e-6;
+      const startCandidates = translatedPoints.filter(
+        (pt) => Math.abs(pt.local.x - minLocalX) < epsilon,
+      );
+      const endCandidates = translatedPoints.filter(
+        (pt) => Math.abs(pt.local.x - maxLocalX) < epsilon,
+      );
+      const averageIso = (candidates) => {
+        if (!candidates.length) {
+          return null;
+        }
+        const sum = candidates.reduce(
+          (acc, pt) => ({ x: acc.x + pt.iso.x, y: acc.y + pt.iso.y }),
+          { x: 0, y: 0 },
+        );
+        return { x: sum.x / candidates.length, y: sum.y / candidates.length };
+      };
+      const isoStart = averageIso(startCandidates);
+      const isoEnd = averageIso(endCandidates);
+      if (isoStart && isoEnd) {
+        const lineLength = Math.hypot(isoEnd.x - isoStart.x, isoEnd.y - isoStart.y);
+        lineSegment = { start: isoStart, end: isoEnd, length: lineLength };
+      }
+    }
+
+    if (lineSegment) {
+      return {
+        index,
+        angle,
+        cos,
+        sin,
+        depth: basePoint.z,
+        facing: cos >= 0,
+        isoBase: translatedIsoBase,
+        points: translatedPoints,
+        pathIsoPoints: [lineSegment.start, lineSegment.end],
+        pathSegmentLengths: [lineSegment.length],
+        pathTotalLength: lineSegment.length,
+        radialIso: lineSegment.end || translatedIsoBase,
+        lineSegment,
+      };
+    }
+
+    return {
+      index,
+      angle,
+      cos,
+      sin,
+      depth: basePoint.z,
+      facing: cos >= 0,
+      isoBase: translatedIsoBase,
+      points: translatedPoints,
+      pathIsoPoints,
+      pathSegmentLengths,
+      pathTotalLength,
+      radialIso: radialTarget?.iso || translatedIsoBase,
+      lineSegment: null,
+    };
+  });
+
+  const bounds = {
+    minX: Number.POSITIVE_INFINITY,
+    minY: Number.POSITIVE_INFINITY,
+    maxX: Number.NEGATIVE_INFINITY,
+    maxY: Number.NEGATIVE_INFINITY,
+  };
+
+  const includePointInBounds = (point) => {
+    if (!point || !Number.isFinite(point.x) || !Number.isFinite(point.y)) {
+      return;
+    }
+    if (point.x < bounds.minX) bounds.minX = point.x;
+    if (point.y < bounds.minY) bounds.minY = point.y;
+    if (point.x > bounds.maxX) bounds.maxX = point.x;
+    if (point.y > bounds.maxY) bounds.maxY = point.y;
+  };
+
+  includePointInBounds(soundIsoPoint);
+  copyInfos.forEach((info) => {
+    includePointInBounds(info.isoBase);
+    includePointInBounds(info.radialIso);
+    info.pathIsoPoints.forEach(includePointInBounds);
+    info.points.forEach((pt) => includePointInBounds(pt.iso));
+    if (info.lineSegment) {
+      includePointInBounds(info.lineSegment.start);
+      includePointInBounds(info.lineSegment.end);
+    }
+  });
+
+  if (
+    Number.isFinite(bounds.minX) &&
+    Number.isFinite(bounds.minY) &&
+    Number.isFinite(bounds.maxX) &&
+    Number.isFinite(bounds.maxY)
+  ) {
+    const currentCenter = {
+      x: (bounds.minX + bounds.maxX) / 2,
+      y: (bounds.minY + bounds.maxY) / 2,
+    };
+    const targetCenter = center;
+    const shift = {
+      x: targetCenter.x - currentCenter.x,
+      y: targetCenter.y - currentCenter.y,
+    };
+
+    if (Math.abs(shift.x) > 1e-6 || Math.abs(shift.y) > 1e-6) {
+      const translatePoint = (point) =>
+        point
+          ? {
+              x: point.x + shift.x,
+              y: point.y + shift.y,
+            }
+          : point;
+
+      soundIsoPoint = translatePoint(soundIsoPoint);
+      copyInfos.forEach((info) => {
+        info.isoBase = translatePoint(info.isoBase);
+        info.radialIso = translatePoint(info.radialIso);
+        info.pathIsoPoints = info.pathIsoPoints.map(translatePoint);
+        info.points = info.points.map((pt) => ({
+          ...pt,
+          iso: translatePoint(pt.iso),
+        }));
+        if (info.lineSegment) {
+          info.lineSegment = {
+            ...info.lineSegment,
+            start: translatePoint(info.lineSegment.start),
+            end: translatePoint(info.lineSegment.end),
+          };
+        }
+      });
+    }
+  }
+
+  const drawInfos = [...copyInfos].sort((a, b) => a.depth - b.depth);
+
+  const drawShape = (info, options = {}) => {
+    const { activeCopyIndex: activeIndex } = options;
+    const isoPoints = info.points.map((pt) => pt.iso);
+    if (!isoPoints.length) {
+      return;
+    }
+    const isActive = info.index === activeIndex;
+    const frontShade = info.facing ? 0.26 : 0.14;
+    const baseAlpha = showFullScene ? frontShade : frontShade + 0.12;
+    const strokeAlpha = isActive ? 0.9 : info.facing ? 0.65 : 0.4;
+    const highlightFirstCopy = info.index === 0;
+    const highlightStrokeStyle = 'rgba(255, 255, 255, 0.23)';
+    const highlightLineWidth = Math.max(0.7, canvas.width * 0.001);
+    if (isLineShape && info.lineSegment) {
+      ctx.save();
+      ctx.strokeStyle = `rgba(60, 47, 87, ${strokeAlpha})`;
+      ctx.lineWidth = Math.max(1.8, canvas.width * (isActive ? 0.0028 : 0.002));
+      ctx.beginPath();
+      ctx.moveTo(info.lineSegment.start.x, info.lineSegment.start.y);
+      ctx.lineTo(info.lineSegment.end.x, info.lineSegment.end.y);
+      ctx.stroke();
+      if (highlightFirstCopy) {
+        ctx.strokeStyle = highlightStrokeStyle;
+        ctx.lineWidth = highlightLineWidth;
+        ctx.stroke();
+      }
+      ctx.restore();
+      return;
+    }
+    if (isCircleShape) {
+      ctx.save();
+      ctx.strokeStyle = `rgba(60, 47, 87, ${strokeAlpha})`;
+      ctx.lineWidth = Math.max(1.8, canvas.width * (isActive ? 0.0026 : 0.0019));
+      ctx.beginPath();
+      ctx.moveTo(isoPoints[0].x, isoPoints[0].y);
+      for (let i = 1; i < isoPoints.length; i += 1) {
+        ctx.lineTo(isoPoints[i].x, isoPoints[i].y);
+      }
+      ctx.closePath();
+      ctx.stroke();
+      if (highlightFirstCopy) {
+        ctx.strokeStyle = highlightStrokeStyle;
+        ctx.lineWidth = highlightLineWidth;
+        ctx.stroke();
+      }
+      ctx.restore();
+      return;
+    }
+    ctx.save();
+    ctx.beginPath();
+    ctx.moveTo(isoPoints[0].x, isoPoints[0].y);
+    for (let i = 1; i < isoPoints.length; i += 1) {
+      ctx.lineTo(isoPoints[i].x, isoPoints[i].y);
+    }
+    ctx.closePath();
+    ctx.fillStyle = `rgba(156, 106, 222, ${Math.min(0.85, baseAlpha + (isActive ? 0.18 : 0))})`;
+    ctx.fill();
+    ctx.strokeStyle = `rgba(60, 47, 87, ${strokeAlpha})`;
+    ctx.lineWidth = Math.max(1.6, canvas.width * (isActive ? 0.0024 : 0.0016));
+    ctx.stroke();
+    if (highlightFirstCopy) {
+      ctx.strokeStyle = highlightStrokeStyle;
+      ctx.lineWidth = highlightLineWidth;
+      ctx.stroke();
+    }
+    ctx.restore();
+  };
+
+  let movingIsoPoint = null;
+  if (copyCycle > 0 && copyInfos[activeCopyIndex]) {
+    const info = copyInfos[activeCopyIndex];
+    const totalLength = info.pathTotalLength;
+    const segmentDuration = view2d.segmentDuration || 0;
+    if (view2d.bounce && info.lineSegment && segmentDuration > 0) {
+      const bounceCycle = segmentDuration * 2;
+      const local = bounceCycle > 0 ? copyElapsed % bounceCycle : 0;
+      const phase = bounceCycle > 0 ? Math.floor(local / segmentDuration) : 0;
+      const t = segmentDuration > 0 ? (local - phase * segmentDuration) / segmentDuration : 0;
+      const fromPoint = phase % 2 === 0 ? info.lineSegment.start : info.lineSegment.end;
+      const toPoint = phase % 2 === 0 ? info.lineSegment.end : info.lineSegment.start;
+      movingIsoPoint = {
+        x: fromPoint.x + (toPoint.x - fromPoint.x) * t,
+        y: fromPoint.y + (toPoint.y - fromPoint.y) * t,
+      };
+    } else if (totalLength > 0) {
+      const progress = copyElapsed / copyCycle;
+      let distance = Math.max(0, Math.min(1, progress)) * totalLength;
+      for (let i = 0; i < info.pathSegmentLengths.length; i += 1) {
+        const length = info.pathSegmentLengths[i];
+        const from = info.pathIsoPoints[i];
+        const to = info.pathIsoPoints[i + 1];
+        if (!(length > 0)) {
+          continue;
+        }
+        if (distance <= length) {
+          const tSegment = distance / length;
+          movingIsoPoint = {
+            x: from.x + (to.x - from.x) * tSegment,
+            y: from.y + (to.y - from.y) * tSegment,
+          };
+          break;
+        }
+        distance -= length;
+      }
+      if (!movingIsoPoint) {
+        const last = info.pathIsoPoints[info.pathIsoPoints.length - 1];
+        movingIsoPoint = last || info.pathIsoPoints[0];
+      }
+    }
+  }
+
+  if (!movingIsoPoint && copyInfos[0]) {
+    movingIsoPoint = copyInfos[0].pathIsoPoints[0];
+  }
+
+  drawInfos.forEach((info) => {
+    const shouldDraw = showFullScene || info.index === activeCopyIndex;
+    if (!shouldDraw) {
+      return;
+    }
+    drawShape(info, { activeCopyIndex });
+  });
+
+  if (movingIsoPoint) {
+    drawMarker(movingIsoPoint, {
+      radius: baseMarkerRadius * 1.1,
+      baseOpacity: 0.32,
+    });
+  }
+
+  drawMarker(soundIsoPoint, {
+    color: 'rgba(255, 255, 255, 0.85)',
+    stroke: 'rgba(255, 255, 255, 0.9)',
+    baseOpacity: 0.18,
+    radius: baseMarkerRadius * 0.75,
+  });
+}
+
+
+
+
+export function drawGatiQuadrant3d({
+  ctx,
+  canvas,
+  orientation,
+  view2d,
+  cycleDuration,
+  palette,
+  elapsed,
+}) {
+  setRenderContext({ ctx, canvas, palette });
+  drawGatiQuadrant3dInternal({ orientation, view2d, cycleDuration }, elapsed);
+}
+
+export function drawJatiQuadrant3d({
+  ctx,
+  canvas,
+  orientation,
+  view2d,
+  cycleDuration,
+  gatiCount,
+  palette,
+  elapsed,
+}) {
+  setRenderContext({ ctx, canvas, palette });
+  drawJatiQuadrant3dInternal({ orientation, view2d, cycleDuration, gatiCount }, elapsed);
+}
+
+export function drawNadaiQuadrant3d({
+  ctx,
+  canvas,
+  orientation,
+  view2d,
+  cycleDuration,
+  gatiCount,
+  palette,
+  elapsed,
+}) {
+  setRenderContext({ ctx, canvas, palette });
+  drawNadaiQuadrant3dInternal({ orientation, view2d, cycleDuration, gatiCount }, elapsed);
+}
